user permissions

+++ base/views.py
@@ -1,5 +1,3 @@
-from multiprocessing import context
-from pickle import NONE
 from django.shortcuts import render, redirect
 from .models import Room, Topic
 from .forms import RoomForm
@@ -8,10 +6,13 @@ from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 
 
 # Create your views here.
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -28,6 +29,8 @@ def loginPage(request):
 
     context = {}
     return render(request, 'base/login_register.html', context)
+
+
 
 def logoutUser(request):
     logout(request)
@@ -46,11 +49,16 @@ def home(request):
     context = { 'rooms' : rooms , 'topics' : topics, 'room_count' : room_count}
     return render(request, 'base/home.html', context)
 
+
+
 def room(request, key):
     room = Room.objects.get(id=key)
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -63,9 +71,14 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+
+@login_required(login_url='login')
 def updateRoom(request, key):
     room = Room.objects.get(id = key)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Forbidden')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance = room)
@@ -76,8 +89,14 @@ def updateRoom(request, key):
     return render(request, 'base/room_form.html', context)
 
 
+
+@login_required(login_url='login')
 def deleteRoom(request, key):
     room = Room.objects.get(id = key)
+
+    if request.user != room.host:
+        return HttpResponse('Forbidden')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
